Refactor UCSC module imports
